+++ internal/autoupdate/connection.go
@@ -123,3 +123,18 @@ func (c *Connection) filter(data map[string]string) {
 		c.histroy[key] = new
 	}
 }
+
+func keysDiff(old []string, new []string) []string {
+	slice := make(map[string]bool, len(old))
+	for _, key := range old {
+		slice[key] = true
+	}
+	added := []string{}
+	for _, key := range new {
+		if slice[key] {
+			continue
+		}
+		added = append(added, key)
+	}
+	return added
+}
